requirements.txt
Django>=5.2,<5.3
gunicorn
python-dotenv
django-rosetta
django-modeltranslation
django-tinymce
django-allauth
django-allauth-ui

Pillow
